+++ src/command.ts
@@ -1,5 +1,4 @@
 import * as vscode from "vscode";
-
 /**
  * 创建终端
  * @returns {vscode.Terminal} 终端实例
@@ -9,15 +8,13 @@ function createTerminal(): vscode.Terminal {
   const terminals = vscode.window.terminals;
   // 查找名为"Runner"的终端，如果不存在则创建
   const shell =
-    terminals.find((item) => item.name === "Runner") ||
+    // terminals.find((item) => item.name === "Runner") ||
     vscode.window.createTerminal("Runner");
 
   // 将终端工作目录设置为当前工作目录
-  shell.sendText(
-    `cd ${vscode.workspace.workspaceFolders?.[0].uri.fsPath}`
-  );
+  shell.sendText(`cd ${vscode.workspace.workspaceFolders?.[0].uri.fsPath}`);
   // shell.sendText('clear');
-  // shell.show(); // 如果需要显示终端，可以取消注释这行代码
+  shell.show(); // 如果需要显示终端，可以取消注释这行代码
   return shell;
 }
 async function inputName(promptText: string): Promise<string | undefined> {
@@ -45,56 +42,87 @@ export async function startProject() {
   let cmd = `django-admin startproject ${projectName}`;
   let shell = createTerminal();
   shell.sendText(cmd);
-
-  // 移动项目文件到当前目录
-  // vscode.window.showWarningMessage(`创建项目成功, 正在移动文件...`);
-  // shell.sendText(`mv ${projectName}/* .`);
-  // // 删除项目文件夹
-  // shell.sendText(`rm ${projectName}`);
+  // TODO 打开项目到当前目录
+  const currentDir = getParams("startproject");
+  if (currentDir) {
+    // 移动项目文件到当前目录
+    // vs.window.showWarningMessage(`创建项目成功, 正在移动文件...`);
+    shell.sendText(`mv ${projectName}/* .`);
+    // // 删除项目文件夹
+    shell.sendText(`rm ${projectName}`);
+  }
   // // 清除终端
   // shell.sendText(`clear`);
-  vscode.window.showInformationMessage(`create project success!`);
+  // vscode.window.showInformationMessage(`create project success!`);
 }
 
 export async function startApp() {
   let appname = await inputName("app");
   let cmd = `python manage.py startapp ${appname}`;
-  let shell = createTerminal();
+  const flag = getParams("startApp");
+  const shell = vscode.window.createTerminal();
   shell.sendText(cmd);
+  if (flag) {
+    let content = `from django.urls import path
+from . import views
+
+app_name='${appname}'
+urlpatterns = [
+    path('', views.index, name='index'),
+]`;
+    shell.sendText(
+      `Set-Content -Path '${appname}/urls.py' -Value "${content}"`
+    );
+    shell.sendText(`New-Item -ItemType Directory -Name "${appname}/templates"`);
+    shell.sendText(`New-Item -ItemType Directory -Name "${appname}/static"`);
+    shell.sendText(`New-Item -ItemType Directory -Name "${appname}/static/img"`);
+    shell.sendText(`New-Item -ItemType Directory -Name "${appname}/static/css"`);
+  }
+
   // shell.sendText(`clear`);
   // 等待终端结果再显示信息
-  vscode.window.showInformationMessage(`create ${appname} success!`);
 }
 
 export function runServer() {
-  const config = vscode.workspace
-    .getConfiguration("django-manager")
-    .get("runServer");
-  const params = JSON.parse(JSON.stringify(config));
+  const port = getParams("runServer");
   let cmd = `python manage.py runserver`;
 
   // let bind = "127.0.0.1";
-  let port = params.port;
+  // let port = params.port;
   // 不为8080
-  if (port && port !== "8080") {
+  if (port && port !== 8080) {
     cmd += ` --port ${port}`;
   }
   let shell = createTerminal();
   shell.sendText(cmd);
   shell.show();
-  vscode.window.showInformationMessage("Django server started");
+  // vscode.window.showInformationMessage("Django server started");
 }
 
 export function makeMigrations() {
   let shell = createTerminal();
   shell.sendText(`python manage.py makemigrations`);
-  // shell.show();
-  vscode.window.showInformationMessage("Migrations created");
+  shell.show();
+  // vscode.window.showInformationMessage("Migrations created");
 }
 
-export function migrate(){
+export function migrate() {
   let shell = createTerminal();
   shell.sendText(`python manage.py migrate`);
-  // shell.show();
-  vscode.window.showInformationMessage("Migrations applied");
+  shell.show();
+  // vscode.window.showInformationMessage("Migrations applied");
+}
+function getParams(configName: string) {
+  const config = vscode.workspace
+    .getConfiguration("django-manager")
+    .get(configName);
+  const params = JSON.parse(JSON.stringify(config));
+  return params;
+}
+
+// 读取文件
+async function readFile(filePath: string) {
+  let fileUri = vscode.Uri.parse(filePath);
+  let content = await vscode.workspace.fs.readFile(fileUri);
+  return content;
 }