+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to the "django-manager" extension will be documented in this
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## [ 0.1.0 ]
+
+- create app with directory (static, templates) and `urls.py`
+- create project in current directory
+
 ## [ 0.0.2 ]
 
 - finish basic commands